+++ src/api/udp/api.go
@@ -3,7 +3,6 @@ package udp
 import (
 	"cluster"
 	. "common"
-	"configuration"
 	"coordinator"
 	"encoding/json"
 	"net"
@@ -22,11 +21,11 @@ type Server struct {
 	shutdown      chan bool
 }
 
-func NewServer(config *configuration.Configuration, coord coordinator.Coordinator, clusterConfig *cluster.ClusterConfiguration) *Server {
+func NewServer(listenAddress string, database string, coord coordinator.Coordinator, clusterConfig *cluster.ClusterConfiguration) *Server {
 	self := &Server{}
 
-	self.listenAddress = config.UdpInputPortString()
-	self.database = config.UdpInputDatabase
+	self.listenAddress = listenAddress
+	self.database = database
 	self.coordinator = coord
 	self.shutdown = make(chan bool, 1)
 	self.clusterConfig = clusterConfig